+++ src/callbacks/liquidity/UniswapV3DTL.sol
@@ -77,7 +77,7 @@ contract UniswapV3DirectToLiquidity is BaseDirectToLiquidity {
         address,
         address baseToken_,
         address quoteToken_,
-        uint96,
+        uint256,
         bool,
         bytes calldata callbackData_
     ) internal virtual override {
